:construction: add uninstall condition in apply.sh

+++ apply.sh
@@ -1,7 +1,19 @@
 #!/bin/zsh
-git submodule init
-git submodule update
-rm -f ~/.zshrc_bk
-mv ~/.zshrc ~/.zshrc_bk
+
+if [ "$1" = "uninstall" ]; then
+    echo "First argument is 'uninstall'"
+	exit 0;
+fi
+
+if [ -d "zplug/.git" ]; then
+    git submodule update --init
+fi
+
+if [ -f "~/.zshrc" ]; then
+    rm -f ~/.zshrc_bk
+    mv ~/.zshrc ~/.zshrc_bk
+fi
+
 ln config/zshrc.zsh ~/.zshrc
+
 source ~/.zshrc
